refactor(frontend): Rename FeeContext to EthFeeContext
