+++ src/frontend/src/eth/components/convert/EthConvertTokenWizard.svelte
@@ -6,7 +6,7 @@
 	import EthConvertForm from '$eth/components/convert/EthConvertForm.svelte';
 	import EthConvertProgress from '$eth/components/convert/EthConvertProgress.svelte';
 	import EthConvertReview from '$eth/components/convert/EthConvertReview.svelte';
-	import FeeContext from '$eth/components/fee/FeeContext.svelte';
+	import EthFeeContext from '$eth/components/fee/EthFeeContext.svelte';
 	import { selectedEthereumNetwork } from '$eth/derived/network.derived';
 	import { ethereumToken } from '$eth/derived/token.derived';
 	import { send as executeSend } from '$eth/services/send.services';
@@ -49,7 +49,7 @@
 
 	const { sourceToken } = getContext<ConvertContext>(CONVERT_CONTEXT_KEY);
 
-	const { feeStore } = getContext<FeeContext>(FEE_CONTEXT_KEY);
+	const { feeStore } = getContext<EthFeeContext>(FEE_CONTEXT_KEY);
 
 	let destination = '';
 	$: destination = isTokenErc20($sourceToken)
@@ -161,7 +161,7 @@
 	const back = () => dispatch('icBack');
 </script>
 
-<FeeContext
+<EthFeeContext
 	sendToken={$sourceToken}
 	sendTokenId={$sourceToken.id}
 	amount={sendAmount}
@@ -196,4 +196,4 @@
 	{:else}
 		<slot />
 	{/if}
-</FeeContext>
+</EthFeeContext>
